created a new datatype test class
