+++ v2/datastream-to-spanner/src/test/resources/DataStreamToSpannerDatatypeIT/mysql-backfill-AllDatatypeColumns.jsonl
