+++ v2/datastream-to-spanner/src/test/resources/DataStreamToSpannerDatatypeIT/mysql-cdc-AllDatatypeColumns.jsonl
